+++ tests/test_client.py
@@ -1,6 +1,6 @@
 import mock
 import httplib2
-from cloudservers.client import CloudServersClient
+from openstack.compute.client import ComputeClient
 from nose.tools import assert_equal
 
 fake_response = httplib2.Response({"status": 200})
@@ -8,7 +8,7 @@ fake_body = '{"hi": "there"}'
 mock_request = mock.Mock(return_value=(fake_response, fake_body))
 
 def client():
-    cl = CloudServersClient("username", "apikey")
+    cl = ComputeClient("username", "apikey")
     cl.management_url = "http://example.com"
     cl.auth_token = "token"
     return cl