+++ tests/test_flavors.py
@@ -1,4 +1,4 @@
-from cloudservers import Flavor, NotFound
+from openstack.compute import Flavor, NotFound
 from fakeserver import FakeServer
 from utils import assert_isinstance
 from nose.tools import assert_raises, assert_equal