+++ tests/test_ipgroups.py
@@ -1,4 +1,4 @@
-from cloudservers import IPGroup
+from openstack.compute import IPGroup
 from fakeserver import FakeServer
 from utils import assert_isinstance
 from nose.tools import assert_equal