+++ tests/test_shell.py
@@ -2,7 +2,7 @@ import os
 import mock
 import httplib2
 from nose.tools import assert_raises, assert_equal
-from cloudservers.shell import CloudserversShell, CommandError
+from openstack.compute.shell import ComputeShell, CommandError
 from fakeserver import FakeServer
 from utils import assert_in
 
@@ -10,15 +10,15 @@ from utils import assert_in
 def setup():
     global _old_env
     fake_env = {
-        'CLOUD_SERVERS_USERNAME': 'username',
-        'CLOUD_SERVERS_API_KEY': 'password'
+        'OPENSTACK_COMPUTE_USERNAME': 'username',
+        'OPENSTACK_COMPUTE_API_KEY': 'password'
     }
     _old_env, os.environ = os.environ, fake_env.copy()
 
     # Make a fake shell object, a helping wrapper to call it, and a quick way
     # of asserting that certain API calls were made.
     global shell, _shell, assert_called
-    _shell = CloudserversShell()
+    _shell = ComputeShell()
     _shell._api_class = FakeServer
     assert_called = lambda m, u, b=None: _shell.cs.assert_called(m, u, b)
     shell = lambda cmd: _shell.main(cmd.split())